feat: move shadcn file installation folder to inside components folder
