+++ packages/lib/src/components/Avatar/Avatar.tsx
@@ -1,3 +1,3 @@
 "use client"
 
-export { Avatar, AvatarFallback, AvatarImage } from "../../_reference/primal/ui/avatar";
+export { Avatar, AvatarFallback, AvatarImage } from "../_base/avatar";
