+++ packages/lib/src/components/Button/Button.tsx
@@ -1,3 +1,3 @@
 "use client"
 
-export { Button } from "../../_reference/primal/ui/button";
+export { Button } from "../_base/button";
