+++ packages/lib/src/components/_base/avatar.tsx
